refactor: separe classes bewteen general css or specific classes + rename Controller
