+++ app/views/home/create-exercise.php
@@ -6,6 +6,7 @@
 <head>
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="/css/style.css" />
+    <link rel="stylesheet" href="/css/create-exercise.css" />
 </head>
 
 <body>
