+++ public/css/create-exercise.css
@@ -1,0 +1,100 @@
+.exercise-label 
+{
+    margin: auto;
+    margin-left: 2rem;
+}
+.managing 
+{
+    background-color: #e0a458;
+    border-color: #e0a458;
+}
+.container
+{
+    margin: auto;
+    max-width: 112.0rem;
+    padding: 0 2.0rem;
+    position: relative;
+    width: 100%;   
+}
+main.container
+{
+    margin-top: 5rem;
+}
+html
+{
+    box-sizing: border-box;
+    font-size: 62.5%;
+}
+div.field, div.actions
+{
+    margin-bottom: 10px;
+}
+header.heading section 
+{
+    display: flex;
+    height: 5rem;
+}
+header.heading 
+{
+    height: 5rem;
+    color: white;
+    display: block;
+    left: 0;
+    max-width: 100%;
+    position: fixed;
+    right: 0;
+    top: 0;
+    width: 100%;
+    z-index: 1;
+}
+header.heading .exercise-label
+{
+    margin: auto;
+    margin-left: 2rem;
+}
+input[type='submit']
+{
+    background-color: #9b4fca;
+    border: 0.1rem solid #9b4fca;
+    border-radius: .4rem;
+    color: #fff;
+    cursor: pointer;
+    display: inline-block;
+    font-size: 1.1rem;
+    font-weight: 700;
+    height: 3.8rem;
+    letter-spacing: .1rem;
+    line-height: 3.8rem;
+    padding: 0 3.0rem;
+    text-align: center;
+    text-decoration: none;
+    text-transform: uppercase;
+    white-space: nowrap;
+}
+input[type='text']
+{
+    -webkit-appearance: none;
+    -moz-appearance: none;
+    appearance: none;
+    background-color: transparent;
+    border: 0.1rem solid #d1d1d1;
+    border-radius: .4rem;
+    box-shadow: none;
+    box-sizing: inherit;
+
+    padding: .6rem 1.0rem;
+    width: 100%;
+}
+div.field, div.actions 
+{
+    margin-bottom: 1opx;
+}
+h1 
+{
+    font-size: 4.6rem;
+    line-height: 1.2;
+    font-weight: 300;
+    letter-spacing: -.1rem;
+    margin-bottom: 2.0rem;
+    margin-top: 0;
+}
